[bugfix] (PLENTY-66) Refactor all route generation occurrences.
