+++ src/Services/PaymentService.php
@@ -104,6 +104,10 @@ class PaymentService
      * @var OrderRepositoryContract
      */
     private $orderRepo;
+    /**
+     * @var UrlServiceContract
+     */
+    private $urlService;
 
     /**
      * PaymentService constructor.
@@ -120,6 +124,7 @@ class PaymentService
      * @param MethodConfigContract $methodConfig
      * @param OrderTxnIdRelationRepositoryContract $orderTxnIdRepo
      * @param OrderRepositoryContract $orderRepo
+     * @param UrlServiceContract $urlService
      */
     public function __construct(
         AddressRepositoryContract $addressRepository,
@@ -133,7 +138,8 @@ class PaymentService
         NotificationServiceContract $notification,
         MethodConfigContract $methodConfig,
         OrderTxnIdRelationRepositoryContract $orderTxnIdRepo,
-        OrderRepositoryContract $orderRepo
+        OrderRepositoryContract $orderRepo,
+        UrlServiceContract $urlService
     ) {
         $this->addressRepository = $addressRepository;
         $this->countryRepository = $countryRepository;
@@ -147,6 +153,7 @@ class PaymentService
         $this->methodConfig = $methodConfig;
         $this->orderTxnIdRepo = $orderTxnIdRepo;
         $this->orderRepo = $orderRepo;
+        $this->urlService = $urlService;
     }
 
     /**
@@ -347,12 +354,12 @@ class PaymentService
 
         $this->heidelpayRequest['FRONTEND_ENABLED']      = 'TRUE';
         $this->heidelpayRequest['FRONTEND_LANGUAGE']     = $this->sessionStorageFactory->getLocaleSettings()->language;
-        $this->heidelpayRequest['FRONTEND_RESPONSE_URL'] = $this->generateUrl(Routes::RESPONSE_URL);
+        $this->heidelpayRequest['FRONTEND_RESPONSE_URL'] = $this->urlService->generateURL(Routes::RESPONSE_URL);
 
         // add the origin domain, which is important for the CSP
         // set 'PREVENT_ASYNC_REDIRECT' to false, to ensure the customer is being redirected after submitting the form.
         if (\in_array($paymentMethod, self::CARD_METHODS, true)) {
-            $this->heidelpayRequest['FRONTEND_PAYMENT_FRAME_ORIGIN'] = $this->paymentHelper->getDomain();
+            $this->heidelpayRequest['FRONTEND_PAYMENT_FRAME_ORIGIN'] = $this->urlService->getDomain();
             $this->heidelpayRequest['FRONTEND_PREVENT_ASYNC_REDIRECT'] = 'false';
         }
 
@@ -373,7 +380,7 @@ class PaymentService
         $this->heidelpayRequest['CRITERION_BASKET_ID'] = $basketArray['id'];
         $this->heidelpayRequest['CRITERION_ORDER_ID'] = $basketArray['orderId'];
         $this->heidelpayRequest['CRITERION_ORDER_TIMESTAMP'] = $basketArray['orderTimestamp'];
-        $this->heidelpayRequest['CRITERION_PUSH_URL'] = $this->generateUrl(Routes::PUSH_NOTIFICATION_URL);
+        $this->heidelpayRequest['CRITERION_PUSH_URL'] = $this->urlService->generateURL(Routes::PUSH_NOTIFICATION_URL);
 
         $secret = $secretService->getSecretHash($transactionId);
         if (null !== $secret) {
@@ -702,19 +709,5 @@ class PaymentService
         $this->orderRepo->updateOrder($order->toArray(), $order->id);
     }
 
-    /**
-     * Generates the full route to the given url.
-     *
-     * @param string $route
-     * @return string
-     */
-    private function generateUrl($route): string
-    {
-        $responseURL = $this->paymentHelper->getDomain() . '/' . $route;
-        if (isset($_COOKIE['PluginSetPreview'])) {
-            $responseURL .= '?pluginSetPreview=' . $_COOKIE ['PluginSetPreview'];
-        }
-        return $responseURL;
-    }
     //</editor-fold>
 }
